test(view): move the selected-scope cases into their own file
view.test.ts carries deliberate credential-redaction fixtures, and the privacy
gate scans whole CHANGED files — so merely appending to it put those pre-existing
fixtures in the publication scope and failed the gate. The new cases live in
selectionScope.test.ts instead, leaving view.test.ts byte-identical to main.
