+++ src/lib/view/view.test.ts
@@ -180,88 +180,6 @@ describe("snapshot text", () => {
   });
 });
 
-describe("selection made OUTSIDE scheme mode reaches the snapshot (#771)", () => {
-  /* The publication half of #771 lives in the client (three publishers reporting
-     one lifted set). This is the consumer half, asserted end to end: given a view
-     session whose mode is `list` or `mobile-focus` and whose selectedPaths are
-     non-empty, the snapshot must return those conversations AND their content.
-     `scopedPaths` and `validateExplicitMembership` already read `selectedPaths`
-     without caring which mode reported them, so nothing in snapshot.ts changed —
-     this pins that down so a future edit cannot quietly make `selected` a
-     scheme-only scope. */
-  function withText(name: string, body: string): string {
-    const pathname = path.join(sandbox, name);
-    fs.writeFileSync(pathname, JSON.stringify({ type: "assistant", timestamp: "t", message: { content: [{ type: "text", text: body }] } }) + "\n");
-    return pathname;
-  }
-
-  test("scope selected returns the selected conversations' content from list mode", async () => {
-    const one = withText("list-selected-one.jsonl", "the first selected conversation");
-    const two = withText("list-selected-two.jsonl", "the second selected conversation");
-    const other = withText("list-unselected.jsonl", "not selected at all");
-    /* A list-mode session: no camera, rows as visiblePaths, and the operator's
-       selection carried in the order the LIST renders. */
-    upsertPresence(presence({ mode: "list", camera: null, focusedPath: null, selectedPaths: [one, two], visiblePaths: [one, two, other] }), 1000);
-
-    const result = await composeSnapshot({
-      request: { schemaVersion: 1, scope: { kind: "selected" } },
-      files: [file(one), file(two), file(other)],
-      siblings: { selfResolution: "omitted", agents: [] },
-      scannerDurationMs: 0,
-      now: 2000,
-    });
-
-    expect(result.view.mode).toBe("list");
-    expect(result.view.selectedPaths).toEqual([one, two]);
-    expect(result.scope).toMatchObject({ kind: "selected", totalPaths: 2, returnedPaths: [one, two], truncated: false });
-    /* Requirement (c): the CONTENT of the selected conversations, not just names. */
-    const bodies = result.conversations.flatMap((item) => item.text?.messages.map((message) => message.text) ?? []);
-    expect(bodies.join("\n")).toContain("the first selected conversation");
-    expect(bodies.join("\n")).toContain("the second selected conversation");
-    expect(bodies.join("\n")).not.toContain("not selected at all");
-  });
-
-  test("focused-selected and explicit paths admit a mobile-focus selection too", async () => {
-    const focused = withText("phone-focused.jsonl", "the pinned pane");
-    const selected = withText("phone-selected.jsonl", "the selected card");
-    upsertPresence(presence({ mode: "mobile-focus", camera: null, focusedPath: focused, selectedPaths: [selected], visiblePaths: [focused] }), 1000);
-
-    const both = await composeSnapshot({
-      request: { schemaVersion: 1, scope: { kind: "focused-selected" }, text: { include: false } },
-      files: [file(focused), file(selected)],
-      siblings: { selfResolution: "omitted", agents: [] },
-      scannerDurationMs: 0,
-      now: 2000,
-    });
-    expect(both.scope.returnedPaths).toEqual([focused, selected]);
-
-    /* A selected path is part of the current view even when this view does not
-       render it, so naming it explicitly is admitted rather than rejected. */
-    const explicit = await composeSnapshot({
-      request: { schemaVersion: 1, scope: { kind: "paths", paths: [selected] }, text: { include: false } },
-      files: [file(focused), file(selected)],
-      siblings: { selfResolution: "omitted", agents: [] },
-      scannerDurationMs: 0,
-      now: 2000,
-    });
-    expect(explicit.scope.returnedPaths).toEqual([selected]);
-  });
-
-  test("an empty selection yields an empty selected scope rather than a fallback", async () => {
-    const only = withText("no-selection.jsonl", "visible but unselected");
-    upsertPresence(presence({ mode: "list", camera: null, focusedPath: null, selectedPaths: [], visiblePaths: [only] }), 1000);
-    const result = await composeSnapshot({
-      request: { schemaVersion: 1, scope: { kind: "selected" }, text: { include: false } },
-      files: [file(only)],
-      siblings: { selfResolution: "omitted", agents: [] },
-      scannerDurationMs: 0,
-      now: 2000,
-    });
-    expect(result.scope).toMatchObject({ kind: "selected", totalPaths: 0, returnedPaths: [] });
-    expect(result.conversations).toEqual([]);
-  });
-});
-
 test("validators bound membership and scope", () => {
   expect(() => validatePresence({ ...presence(), visiblePaths: Array.from({ length: 129 }, (_, index) => `/p-${index}`) })).toThrow(ViewValidationError);
   expect(() => validateSnapshotRequest({ schemaVersion: 1, scope: { kind: "paths", paths: Array.from({ length: 17 }, (_, index) => `/p-${index}`) } })).toThrow(ViewValidationError);
